src/test/java/club/yaso91/alarm_server/controller/AlarmInfoControllerTest.java
package club.yaso91.alarm_server.controller;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.ResultActions;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.Assert.assertEquals;

@SpringBootTest
@RunWith(SpringRunner.class)
@Transactional
public class AlarmInfoControllerTest {
    @Autowired
    private AlarmInfoController alarmInfoController;
    private MockMvc mockMvc;

    @Before
    public void setUp() throws Exception {
        mockMvc = MockMvcBuilders.standaloneSetup(alarmInfoController).build();
    }


    @Test
    public void getAlarmInfos() throws Exception {
        RequestBuilder requestBuilder = MockMvcRequestBuilders.get("/alarmInfo")
                .contentType(MediaType.APPLICATION_JSON_UTF8);
        ResultActions perform = mockMvc.perform(requestBuilder);

        MvcResult result = perform.andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.handler().methodName("getAlarmInfos"))
                .andDo(MockMvcResultHandlers.print())
                .andReturn();
    }
}
